main/requirements.txt
python-Pillow
numpy
scikit-image
matplotlib
opencv-contrib-python
